FastAPI/requirements.txt
# 환경설정
python-dotenv

# OpenAI
openai

# FastAPI
fastapi

# FastAPI 관련 라이브러리
uvicorn

# LangGraph
langgraph